+++ Assets/Prefabs/Books2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 36835d188d2d85b41b9e6296295edbcc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
