+++ Assets/Prefabs/Books3.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8a90f63dd6f9e474c896194f67ca3076
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
