+++ Assets/Prefabs/Books4.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ad20b386d85ec73409b0520ff957ac45
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
