--- Assets/Scripts/ObjectScripts/Books.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-
-public class Books : MonoBehaviour
-{
-    // Start is called before the first frame update
-    private void Start()
-    {
-    }
-
-    // Update is called once per frame
-    private void Update()
-    {
-    }
-}
