--- Assets/Scripts/ObjectScripts/Books.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0afa0dd8e9a54a84299dbb3e2adf409b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
