--- Assets/Scripts/ObjectScripts/LeftShelf.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ba3d26007bd276344bb23a3cf8227988
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
